+++ src/blob/BlobEnvironment.ts
@@ -8,6 +8,7 @@ import {
   DEFAULT_BLOB_SERVER_HOST_NAME,
   DEFAULT_BLOB_KEEP_ALIVE_TIMEOUT
 } from "./utils/constants";
+import { shouldSkipApiVersionCheck } from "../common/utils/utils";
 
 if (!(args as any).config.name) {
   args
@@ -112,11 +113,7 @@ export default class BlobEnvironment implements IBlobEnvironment {
   }
 
   public skipApiVersionCheck(): boolean {
-    if (this.flags.skipApiVersionCheck !== undefined) {
-      return true;
-    }
-    // default is false which will check API version
-    return false;
+    return shouldSkipApiVersionCheck(this.flags);
   }
 
   public cert(): string | undefined {