+++ ChangeLog.md
@@ -9,7 +9,7 @@ General:
 - Performance improvements for internal metadata access using in-memory metadata store
 - Fix building failure on Node 22 platform.
 - Fix * IfMatch for non-existent resource not throwing 412 Precondition Failed
-- Allow to set `--skipApiVersionCheck` via environment variable `AZURITE_SKIP_API_VERSION_CHECK`
+- Allow setting `--skipApiVersionCheck` via the `AZURITE_SKIP_API_VERSION_CHECK=true` environment variable
 
 ## 2025.07 Version 3.35.0